+++ 1942 Bootstrap/Game1942/SmallShip.h
@@ -5,7 +5,7 @@ class SmallShip : public Enemy
 {
 private:
 	int temp	  = 0;
-	float respwan = rand() % 1280 + 1;
+	int respwan = rand() % 1280 + 1;
 public:
 	SmallShip();
 	void Move(float deltaTime, int sW, int sH) override;
